// src/types/genre.ts
export type Genre =
  | 'боевик'
  | 'приключения'
  | 'мультфильм'
  | 'биография'
  | 'комедия'
  | 'криминал'
  | 'драма'
  | 'документальный'
  | 'семейный'
  | 'фэнтези'
  | 'исторический'
  | 'ужасы'
  | 'музыка'
  | 'мюзикл'
  | 'мелодрама'
  | 'научный'
  | 'спорт'
  | 'триллер'
  | 'военный'
  | 'вестерн'
  | 'детектив'
  | 'анимация'
  | 'фантастика';

export type OptionSelectGenre = {
  value: Genre;
  label: string;
};